makaobank/src/main/java/kr/megaptera/makaobank/services/TransferService.java
package kr.megaptera.makaobank.services;

import kr.megaptera.makaobank.exceptions.*;
import kr.megaptera.makaobank.models.*;
import kr.megaptera.makaobank.repositories.*;
import org.springframework.stereotype.*;
import org.springframework.transaction.annotation.*;

@Service
@Transactional
public class TransferService {

  private AccountRepository accountRepository;

  public TransferService(AccountRepository accountRepository) {
    this.accountRepository = accountRepository;
  }

  public Long transfer(String from, String to, Long amount) {
    //TODO. amount가 0 이하이면 안되고, 내 잔액보다 커서도 안된다.

    Account account1 = accountRepository.findByAccountNumber(from)
        .orElseThrow(() -> new AccountNotFound(from));
    Account account2 = accountRepository.findByAccountNumber(to)
        .orElseThrow(() -> new AccountNotFound(to));

    account1.transfer(account2, amount);

    return amount;
  }
}
